+++ resources/views/admin/partials/menu.blade.php
@@ -15,7 +15,7 @@
     <tr><td><a href="{{ route('admin.resource.subtract') }}" accesskey="k">@lang('admin/menu.resources-sub')</a></td></tr>
     <tr><td class="separated"><a href="colonies.php" accesskey="1">@lang('admin/menu.planets')</a></td></tr>
     <tr><td><a href="activeplanet.php" accesskey="k">@lang('admin/menu.planets-active')</a></td></tr>
-    <tr><td><a href="colonies.php?type=moons" accesskey="k">@lang('admin/menu.moons')</a></td></tr>
+    <tr><td><a href="{{ route('admin.moons') }}" accesskey="k">@lang('admin/menu.moons')</a></td></tr>
     <tr><td><a href="add_moon.php" accesskey="k">@lang('admin/menu.moon-add')</a></td></tr>
     <tr><td><a href="ShowFlyingFleets.php" accesskey="k">@lang('admin/menu.fleets')</a></td></tr>
     <tr><td class="separated"><a href="{{ route('admin.ban') }}" accesskey="k">@lang('admin/menu.ban')</a></td></tr>