+++ CHANGELOG.md
@@ -12,6 +12,7 @@ changes made when rewriting.
 
 - Admin: Ban
     - Translated to English
+- Admin: List of moons
 - Admin: Unban
     - Translated to English
 - Banned